test: fix short-wait-time imports
