+++ __tests__/utils.test.ts
@@ -1,6 +1,4 @@
 import {iterateWithNext} from '@atb/utils/array';
-import {addMinutes} from 'date-fns';
-import {Leg} from '@atb/api/types/trips';
 import {Flattened, flattenObject} from '@atb/utils/object';
 import {onlyUniques, onlyUniquesBasedOnField} from '@atb/utils/only-uniques';
 import {compareVersion} from '@atb/utils/compare-version';
@@ -8,10 +6,6 @@ import {
   expectNumber,
   expectStringEqual,
 } from '../e2e/test/utils/jestAssertions';
-import {
-  hasShortWaitTime,
-  TIME_LIMIT_IN_MINUTES,
-} from '@atb/travel-details-screens/utils';
 import {
   formatToClock,
   formatToClockOrRelativeMinutes,
@@ -47,59 +41,6 @@ describe('IterateWithNext', () => {
       {current: 1, next: 2},
       {current: 2, next: 3},
     ]);
-  });
-});
-
-describe('Short wait time evaluator', () => {
-  const nowDate = Date.now();
-  const Leg1: Leg = {
-    expectedStartTime: nowDate,
-    expectedEndTime: addMinutes(nowDate, 5),
-  } as Leg;
-
-  const Leg2: Leg = {
-    expectedStartTime: addMinutes(nowDate, 5 + TIME_LIMIT_IN_MINUTES - 1),
-    expectedEndTime: addMinutes(nowDate, 10),
-  } as Leg;
-
-  const Leg3: Leg = {
-    expectedStartTime: addMinutes(nowDate, 5 + TIME_LIMIT_IN_MINUTES + 1),
-    expectedEndTime: addMinutes(nowDate, 10),
-  } as Leg;
-
-  const Leg4: Leg = {
-    expectedStartTime: addMinutes(nowDate, 15),
-    expectedEndTime: addMinutes(nowDate, 20),
-  } as Leg;
-
-  const weirdLeg: Leg = {
-    expectedStartTime: 'non parsable string',
-    expectedEndTime: {weird: true},
-  } as Leg;
-
-  it('catches a short wait', () => {
-    const isShortWait = hasShortWaitTime([Leg1, Leg2]);
-    expect(isShortWait).toBe(true);
-  });
-  it('passes on a long wait', () => {
-    const isShortWait = hasShortWaitTime([Leg1, Leg3]);
-    expect(isShortWait).toBe(false);
-  });
-  it('catches a short wait with more legs', () => {
-    const isShortWait = hasShortWaitTime([Leg1, Leg2, Leg4]);
-    expect(isShortWait).toBe(true);
-  });
-  it('passes on only one leg', () => {
-    const isShortWait = hasShortWaitTime([Leg1]);
-    expect(isShortWait).toBe(false);
-  });
-  it('passes with empty array', () => {
-    const isShortWait = hasShortWaitTime([]);
-    expect(isShortWait).toBe(false);
-  });
-  it('passes on weird data', () => {
-    const isShortWait = hasShortWaitTime([weirdLeg, weirdLeg]);
-    expect(isShortWait).toBe(false);
   });
 });
